haritada zoom yapma özelliği

+++ src/entity/Entity.java
@@ -4,9 +4,11 @@ import java.awt.image.BufferedImage;
 
 public class Entity {
 
-	
-	public int worldX, worldY;
-	public int speed;
+	//
+	//  map'te zoom yapabilmek için double haline getiriyoruz.
+	//
+	public double worldX, worldY;
+	public double speed;
 	
 	public BufferedImage up1, up2 ,up3, left1, left2, left3, right1, right2, right3, down1 ,down2 ,down3;
 	public String direction;
